refactor: densify leftover data tests and ratchet the 250-line cap (#4626)
Split remaining over-cap suites by concern and drop repo.data_lib_file_size / repo.data_test_file_size from 300/320 to 250 so wave-7 Slice D can land without a new Quality job.
