+++ SPEC/program/data-test-file-size.md
@@ -25,16 +25,16 @@ Support helpers under `test/support/` share the same ceiling.
 
 ## Ceiling
 
-**320 physical lines** per file (empty shrink-only grandfather allowlist).
+**250 physical lines** per file (empty shrink-only grandfather allowlist).
 
 ## Acceptance criteria
 
 - Given every `.dart` file under `packages/colonizethis_data/test/**` is at or
-  below 320 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule
-  `repo.data_test_file_size`, then the rule passes and exits `0` (Refs #4121, #4292, #4412).
-- Given a data test file exceeds 320 physical lines and is not on the
+  below 250 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule
+  `repo.data_test_file_size`, then the rule passes and exits `0` (Refs #4121, #4292, #4412, #4626).
+- Given a data test file exceeds 250 physical lines and is not on the
   grandfather allowlist, when the checker runs, then the run fails listing the
-  offending path and exits `1` (Refs #4121, #4412).
+  offending path and exits `1` (Refs #4121, #4412, #4626).
 - Given a stale grandfather allowlist entry (missing file), when the checker
   runs, then the run fails with a stale-entry message and exits `1` (Refs
   #4121).
