+++ packages/colonizethis_data/test/ai_parameter_registry_test.dart
@@ -143,132 +143,6 @@ void main() {
       }
     });
   });
-
-  group('AiParameter metadata', () {
-    test('every parameter has populated, ordered, described metadata', () {
-      for (final p in AiParameterRegistry.allParams) {
-        expect(p.name, isNotEmpty);
-        expect(p.description.trim(), isNotEmpty, reason: p.name);
-        expect(p.category, isNotEmpty, reason: p.name);
-        expect(
-          p.minValue <= p.defaultValue,
-          isTrue,
-          reason: '${p.name}: min ${p.minValue} > default ${p.defaultValue}',
-        );
-        expect(
-          p.defaultValue <= p.maxValue,
-          isTrue,
-          reason: '${p.name}: default ${p.defaultValue} > max ${p.maxValue}',
-        );
-      }
-    });
-
-    test('uses one of the four known categories', () {
-      const categories = {
-        AiParameterCategory.personalityDomain,
-        AiParameterCategory.personalityGoal,
-        AiParameterCategory.personalityThreshold,
-        AiParameterCategory.victoryConfig,
-      };
-      for (final p in AiParameterRegistry.allParams) {
-        expect(categories, contains(p.category), reason: p.name);
-      }
-    });
-  });
-
-  group('Canonical key naming', () {
-    test('personality keys are <sourceMap>.<field>', () {
-      final personality = [
-        ...AiParameterRegistry.byCategory(
-          AiParameterCategory.personalityDomain,
-        ),
-        ...AiParameterRegistry.byCategory(AiParameterCategory.personalityGoal),
-        ...AiParameterRegistry.byCategory(
-          AiParameterCategory.personalityThreshold,
-        ),
-      ];
-      expect(personality, isNotEmpty);
-      for (final p in personality) {
-        expect(p.name, contains('.'), reason: p.name);
-        expect(p.name, isNot(startsWith('k')), reason: p.name);
-      }
-    });
-
-    test('victory-config keys are flat Dart constant identifiers', () {
-      final victory = AiParameterRegistry.byCategory(
-        AiParameterCategory.victoryConfig,
-      );
-      expect(victory, isNotEmpty);
-      for (final p in victory) {
-        expect(p.name, startsWith('k'), reason: p.name);
-        expect(p.name, isNot(contains('.')), reason: p.name);
-      }
-    });
-  });
-
-  group('Bounds rules', () {
-    test('personality params are bounded [0, 100]', () {
-      final personality = [
-        ...AiParameterRegistry.byCategory(
-          AiParameterCategory.personalityDomain,
-        ),
-        ...AiParameterRegistry.byCategory(AiParameterCategory.personalityGoal),
-        ...AiParameterRegistry.byCategory(
-          AiParameterCategory.personalityThreshold,
-        ),
-      ];
-      for (final p in personality) {
-        expect(p.isInteger, isTrue, reason: p.name);
-        expect(p.minValue, 0, reason: p.name);
-        expect(p.maxValue, 100, reason: p.name);
-      }
-    });
-
-    test('victory-config int bounds are [0, max(2000, 4x default)]', () {
-      final victoryInts = AiParameterRegistry.byCategory(
-        AiParameterCategory.victoryConfig,
-      ).where((p) => p.isInteger);
-      expect(victoryInts, isNotEmpty);
-      for (final p in victoryInts) {
-        expect(p.minValue, 0, reason: p.name);
-        expect(p.maxValue, math.max(2000, 4 * p.defaultValue), reason: p.name);
-      }
-    });
-
-    test('victory-config double bounds are [0.0, 4x default]', () {
-      final victoryDoubles = AiParameterRegistry.byCategory(
-        AiParameterCategory.victoryConfig,
-      ).where((p) => !p.isInteger);
-      expect(victoryDoubles, isNotEmpty);
-      for (final p in victoryDoubles) {
-        expect(p.minValue, 0.0, reason: p.name);
-        expect(p.maxValue, 4 * p.defaultValue, reason: p.name);
-      }
-    });
-  });
-
-  group('Registry lookups', () {
-    test('byName returns the matching parameter or null', () {
-      final p = AiParameterRegistry.byName('kDeclareWarAdjacentOwnerBonus');
-      expect(p, isNotNull);
-      expect(p!.category, AiParameterCategory.victoryConfig);
-      expect(AiParameterRegistry.byName('not_a_real_parameter'), isNull);
-    });
-
-    test('byCategory partitions allParams exactly', () {
-      final sumByCategory =
-          [
-            AiParameterCategory.personalityDomain,
-            AiParameterCategory.personalityGoal,
-            AiParameterCategory.personalityThreshold,
-            AiParameterCategory.victoryConfig,
-          ].fold<int>(
-            0,
-            (acc, c) => acc + AiParameterRegistry.byCategory(c).length,
-          );
-      expect(sumByCategory, AiParameterRegistry.allParams.length);
-    });
-  });
 }
 
 /// In-scope `const int` / `const double` `k*` names from victory-config modules.
