+++ packages/colonizethis_data/test/support/combat_config_spec_regiments.dart
@@ -1,282 +1,19 @@
 // SPEC/game/military-units.md regiment table fixtures for combat_config_test.dart
-// (Refs #4121 slice D).
-import 'package:colonizethis_data/colonizethis_data.dart';
+// (Refs #4121 slice D, #4626 slice D era split).
+import 'combat_config_spec_regiment_row.dart';
+import 'combat_config_spec_regiments_era1.dart';
+import 'combat_config_spec_regiments_era2.dart';
+import 'combat_config_spec_regiments_era3.dart';
+import 'combat_config_spec_regiments_era4.dart';
 
-/// Expected tactical row per SPEC/game/military-units.md § Regiment Table.
-typedef CombatConfigSpecRegimentRow = ({
-  int fpn,
-  int fpm,
-  int rng,
-  int def,
-  int mvr,
-  RegimentCategory category,
-  int era,
-});
+export 'combat_config_spec_regiment_row.dart';
 
 /// Keys: regiment id. Values: FPN, FPM, RNG, DEF, MVR, category, era.
-const combatConfigMilitaryUnitsSpecById =
-    <String, CombatConfigSpecRegimentRow>{
-  'peasant_levies': (
-    fpn: 0,
-    fpm: 3,
-    rng: 1,
-    def: 3,
-    mvr: 3,
-    category: RegimentCategory.lightInfantry,
-    era: 1,
-  ),
-  'pikemen': (
-    fpn: 0,
-    fpm: 5,
-    rng: 1,
-    def: 5,
-    mvr: 3,
-    category: RegimentCategory.regularInfantry,
-    era: 1,
-  ),
-  'arquebusiers': (
-    fpn: 5,
-    fpm: 1,
-    rng: 3,
-    def: 3,
-    mvr: 2,
-    category: RegimentCategory.heavyInfantry,
-    era: 1,
-  ),
-  'bowmen': (
-    fpn: 3,
-    fpm: 1,
-    rng: 4,
-    def: 2,
-    mvr: 3,
-    category: RegimentCategory.bowmen,
-    era: 1,
-  ),
-  'squires': (
-    fpn: 0,
-    fpm: 4,
-    rng: 1,
-    def: 4,
-    mvr: 6,
-    category: RegimentCategory.lightCavalry,
-    era: 1,
-  ),
-  'knights': (
-    fpn: 0,
-    fpm: 6,
-    rng: 1,
-    def: 6,
-    mvr: 4,
-    category: RegimentCategory.spearCavalry,
-    era: 1,
-  ),
-  'culverin': (
-    fpn: 8,
-    fpm: 1,
-    rng: 5,
-    def: 2,
-    mvr: 2,
-    category: RegimentCategory.heavyArtillery,
-    era: 1,
-  ),
-  'calivermen': (
-    fpn: 3,
-    fpm: 2,
-    rng: 5,
-    def: 5,
-    mvr: 4,
-    category: RegimentCategory.lightInfantry,
-    era: 2,
-  ),
-  'halberdiers': (
-    fpn: 0,
-    fpm: 7,
-    rng: 1,
-    def: 6,
-    mvr: 4,
-    category: RegimentCategory.regularInfantry,
-    era: 2,
-  ),
-  'musketeers': (
-    fpn: 7,
-    fpm: 2,
-    rng: 4,
-    def: 4,
-    mvr: 3,
-    category: RegimentCategory.heavyInfantry,
-    era: 2,
-  ),
-  'cossacks': (
-    fpn: 0,
-    fpm: 5,
-    rng: 1,
-    def: 5,
-    mvr: 8,
-    category: RegimentCategory.lightCavalry,
-    era: 2,
-  ),
-  'lancers': (
-    fpn: 0,
-    fpm: 8,
-    rng: 1,
-    def: 5,
-    mvr: 6,
-    category: RegimentCategory.spearCavalry,
-    era: 2,
-  ),
-  'harquebusiers': (
-    fpn: 2,
-    fpm: 6,
-    rng: 3,
-    def: 5,
-    mvr: 6,
-    category: RegimentCategory.heavyCavalry,
-    era: 2,
-  ),
-  'horse_artillery': (
-    fpn: 5,
-    fpm: 2,
-    rng: 7,
-    def: 2,
-    mvr: 3,
-    category: RegimentCategory.lightArtillery,
-    era: 2,
-  ),
-  'royal_artillery': (
-    fpn: 9,
-    fpm: 2,
-    rng: 8,
-    def: 2,
-    mvr: 2,
-    category: RegimentCategory.heavyArtillery,
-    era: 2,
-  ),
-  'skirmishers': (
-    fpn: 4,
-    fpm: 3,
-    rng: 5,
-    def: 6,
-    mvr: 6,
-    category: RegimentCategory.lightInfantry,
-    era: 3,
-  ),
-  'regulars': (
-    fpn: 7,
-    fpm: 7,
-    rng: 5,
-    def: 5,
-    mvr: 4,
-    category: RegimentCategory.regularInfantry,
-    era: 3,
-  ),
-  'grenadiers': (
-    fpn: 10,
-    fpm: 8,
-    rng: 5,
-    def: 5,
-    mvr: 4,
-    category: RegimentCategory.heavyInfantry,
-    era: 3,
-  ),
-  'hussars': (
-    fpn: 2,
-    fpm: 8,
-    rng: 3,
-    def: 6,
-    mvr: 11,
-    category: RegimentCategory.lightCavalry,
-    era: 3,
-  ),
-  'cuirassiers': (
-    fpn: 5,
-    fpm: 13,
-    rng: 3,
-    def: 5,
-    mvr: 9,
-    category: RegimentCategory.heavyCavalry,
-    era: 3,
-  ),
-  'light_artillery': (
-    fpn: 8,
-    fpm: 3,
-    rng: 9,
-    def: 3,
-    mvr: 4,
-    category: RegimentCategory.lightArtillery,
-    era: 3,
-  ),
-  'heavy_artillery': (
-    fpn: 13,
-    fpm: 2,
-    rng: 10,
-    def: 2,
-    mvr: 3,
-    category: RegimentCategory.heavyArtillery,
-    era: 3,
-  ),
-  'sharpshooters': (
-    fpn: 5,
-    fpm: 4,
-    rng: 7,
-    def: 7,
-    mvr: 7,
-    category: RegimentCategory.lightInfantry,
-    era: 4,
-  ),
-  'rifle_infantry': (
-    fpn: 9,
-    fpm: 9,
-    rng: 6,
-    def: 6,
-    mvr: 4,
-    category: RegimentCategory.regularInfantry,
-    era: 4,
-  ),
-  'guards': (
-    fpn: 12,
-    fpm: 10,
-    rng: 6,
-    def: 6,
-    mvr: 4,
-    category: RegimentCategory.heavyInfantry,
-    era: 4,
-  ),
-  'scouts': (
-    fpn: 5,
-    fpm: 11,
-    rng: 5,
-    def: 6,
-    mvr: 11,
-    category: RegimentCategory.lightCavalry,
-    era: 4,
-  ),
-  'carbine_cavalry': (
-    fpn: 7,
-    fpm: 17,
-    rng: 5,
-    def: 5,
-    mvr: 9,
-    category: RegimentCategory.heavyCavalry,
-    era: 4,
-  ),
-  'field_artillery': (
-    fpn: 10,
-    fpm: 3,
-    rng: 11,
-    def: 4,
-    mvr: 5,
-    category: RegimentCategory.lightArtillery,
-    era: 4,
-  ),
-  'siege_guns': (
-    fpn: 17,
-    fpm: 2,
-    rng: 12,
-    def: 3,
-    mvr: 3,
-    category: RegimentCategory.heavyArtillery,
-    era: 4,
-  ),
+const combatConfigMilitaryUnitsSpecById = <String, CombatConfigSpecRegimentRow>{
+  ...combatConfigMilitaryUnitsSpecByIdEra1,
+  ...combatConfigMilitaryUnitsSpecByIdEra2,
+  ...combatConfigMilitaryUnitsSpecByIdEra3,
+  ...combatConfigMilitaryUnitsSpecByIdEra4,
 };
 
 /// Table order in SPEC/game/military-units.md (must match [regimentCatalog] order).
