+++ packages/colonizethis_data/test/support/tech_graph_spec_ids.dart
@@ -1,0 +1,118 @@
+import 'package:colonizethis_data/colonizethis_data.dart';
+
+/// SPEC/game tech-tree.md ids via kTechId* (Refs #4626 Slice D).
+const Set<String> techGraphSpecTechIds = {
+  kTechIdAdvancedHullDesign,
+  kTechIdAdvancedIronWorking,
+  kTechIdAmalgamationProcess,
+  kTechIdAnimalHusbandry,
+  kTechIdApprenticeWorkers,
+  kTechIdBanking,
+  kTechIdBayonet,
+  kTechIdCigarProduction,
+  kTechIdCircularSaw,
+  kTechIdClipperShips,
+  kTechIdCoalMining,
+  kTechIdConvoying,
+  kTechIdCopperAndTinMining,
+  kTechIdCottonGin,
+  kTechIdCottonPlanting,
+  kTechIdCottonWeaving,
+  kTechIdCropRotation,
+  kTechIdCrucibleProcess,
+  kTechIdDiplomaticExpertise,
+  kTechIdDiscoveryOfCotton,
+  kTechIdDiscoveryOfFurs,
+  kTechIdDiscoveryOfGemsOrDiamonds,
+  kTechIdDiscoveryOfGoldOrSilver,
+  kTechIdDiscoveryOfSpices,
+  kTechIdDiscoveryOfSugar,
+  kTechIdDiscoveryOfTobacco,
+  kTechIdDynamite,
+  kTechIdEarlyRifles,
+  kTechIdEarlySteamEngine,
+  kTechIdEfficientExtractionOfCopperAndTin,
+  kTechIdEliteMilitaryTraining,
+  kTechIdEmpireBuilding,
+  kTechIdEmplacedSiegeGuns,
+  kTechIdExcessiveFurHarvesting,
+  kTechIdExplosives,
+  kTechIdExtractionOfPreciousMetals,
+  kTechIdFieldArtilleryTactics,
+  kTechIdGeologicalProspecting,
+  kTechIdHatProduction,
+  kTechIdHeavyArtillery,
+  kTechIdHeavyEmplacedArtillery,
+  kTechIdHighGradeSteel,
+  kTechIdHorseArtillery,
+  kTechIdHussars,
+  kTechIdImprovedCavalryTactics,
+  kTechIdImprovedCavalryWeapons,
+  kTechIdImprovedFoodPreservation,
+  kTechIdImprovedInfantryTactics,
+  kTechIdImprovedIronWeapons,
+  kTechIdImprovedSailDesign,
+  kTechIdImprovedSeaRoutes,
+  kTechIdImprovedTrappingTechniques,
+  kTechIdIndustrialFundingOfResearch,
+  kTechIdIndustrialIronMining,
+  kTechIdIndustrialMachinery,
+  kTechIdIronMining,
+  kTechIdLandEnclosure,
+  kTechIdLargeCoalMines,
+  kTechIdLargeCopperAndTinMines,
+  kTechIdLargeCottonPlantations,
+  kTechIdLargeHulls,
+  kTechIdLargePreciousStoneMines,
+  kTechIdLargeSpicePlantations,
+  kTechIdLargeSugarPlantations,
+  kTechIdLargeTobaccoPlantations,
+  kTechIdLaterSteamEngine,
+  kTechIdLightArtilleryTactics,
+  kTechIdLongRangeRifles,
+  kTechIdMasterArtisans,
+  kTechIdMerchantCompanies,
+  kTechIdMerchantSteamships,
+  kTechIdMineEngineering,
+  kTechIdModernForts,
+  kTechIdModernMilitaryFunding,
+  kTechIdMoldboardPlow,
+  kTechIdMoneyLending,
+  kTechIdNationalBureaucracy,
+  kTechIdNationalism,
+  kTechIdNavigation,
+  kTechIdNeedleGuns,
+  kTechIdOrganisedRegiments,
+  kTechIdPaddlewheels,
+  kTechIdPreciousMetalsMining,
+  kTechIdPreciousStoneMining,
+  kTechIdPrintingPress,
+  kTechIdPrivateeringCompanies,
+  kTechIdPropaganda,
+  kTechIdRecruitSteppeHorsemen,
+  kTechIdRepeatingCavalryCarbine,
+  kTechIdRiverboats,
+  kTechIdRoadConstruction,
+  kTechIdSafetyLamp,
+  kTechIdSawMill,
+  kTechIdScientificCattleBreeding,
+  kTechIdScientificSheepBreeding,
+  kTechIdScouting,
+  kTechIdSeedDrill,
+  kTechIdSheepRanching,
+  kTechIdShipOfTheLine,
+  kTechIdSiegeEngineering,
+  kTechIdSquareSetTimbering,
+  kTechIdSteamInMining,
+  kTechIdSugarIndustry,
+  kTechIdSugarPlanting,
+  kTechIdSugarRefining,
+  kTechIdSuperiorHullDesign,
+  kTechIdTobaccoIndustry,
+  kTechIdTobaccoPlanting,
+  kTechIdTradeFairs,
+  kTechIdTrainedJourneymen,
+  kTechIdUniversity,
+  kTechIdWeaponCraftsmanship,
+  kTechIdWindSawMill,
+};
