+++ packages/colonizethis_data/test/support/work_order_target_ids.dart
@@ -1,0 +1,10 @@
+const kWorkTargetExplore = 'explore';
+const kWorkTargetProspect = 'prospect';
+const kWorkTargetBuildImprovement = 'build_improvement';
+const kWorkTargetUpgradeTown = 'upgrade_town';
+const kWorkTargetBuildRoad = 'build_road';
+const kWorkTargetBuildPort = 'build_port';
+const kWorkTargetBuildFort = 'build_fort';
+const kWorkTargetBuildRail = 'build_rail';
+const kWorkTargetCounterSpy = 'counter_spy';
+const kWorkTargetPurchaseLand = 'purchase_land';
